+++ package.json
@@ -10,6 +10,8 @@
     "@fortawesome/fontawesome-free": "^6.1.1",
     "bulma": "^0.9.4",
     "classnames": "^2.3.1",
+    "lodash": "^4.17.21",
+    "lodash.debounce": "^4.0.8",
     "postcss": "^8.4.12",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -23,6 +25,7 @@
     "@mate-academy/scripts": "^1.2.8",
     "@mate-academy/students-ts-config": "*",
     "@mate-academy/stylelint-config": "*",
+    "@types/lodash.debounce": "^4.0.7",
     "@types/node": "^17.0.23",
     "@types/react": "^17.0.43",
     "@types/react-dom": "^17.0.14",
